Provide both Fedora and RHEL files in the SRPM
Currently, we've been switching between copying Fedora and RHEL files
based on dist-tag. Long term, we want to have a single tree to manage
both Fedora and RHEL. Move towards this by shipping both RHEL and
Fedora files in a single RPM. The choice of which set of files to use
gets decided at build time based on the buildroot.

This series is a big large but it's not easy to break up simply because
making one change requires making all the other changes to avoid breaking
anything. The big changes:

- Move some files and scripts to have the .fedora or .rhel extension
- Tweak some of the config scripts to accont for a fedora/rhel extension
- spec file changes to account for these changes

Signed-off-by: Laura Abbott <labbott@redhat.com>