[redhat] New configs in drivers/thunderbolt
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_USB4:

 USB4 and Thunderbolt driver. USB4 is the public speficiation
 based on Thunderbolt 3 protocol. This driver is required if
 you want to hotplug Thunderbolt and USB4 compliant devices on
 Apple hardware or on PCs with Intel Falcon Ridge or newer.

 To compile this driver a module, choose M here. The module will be
 called thunderbolt.

 Symbol: USB4 [=n]
 Type  : tristate
 Defined at drivers/thunderbolt/Kconfig:2
   Prompt: Unified support for USB4 and Thunderbolt
   Depends on: PCI [=y] && (X86 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
 Selects: APPLE_PROPERTIES [=n] && CRC32 [=y] && CRYPTO [=y] && CRYPTO_HASH [=y] && NVMEM [=y]

---

Jarod requested that CONFIG_USB4=y since it's a direct replacement for
CONFIG_THUNDERBOLT which is currently y. This has exposed a further
configuration:

 CONFIG_USB4_NET:

 Select this if you want to create network between two computers
 over a USB4 and Thunderbolt cables. The driver supports Apple
 ThunderboltIP protocol and allows communication with any host
 supporting the same protocol including Windows and macOS.

 To compile this driver a module, choose M here. The module will be
 called thunderbolt-net.

 Symbol: USB4_NET [=n]
 Type  : tristate
 Defined at drivers/net/Kconfig:535
   Prompt: Networking over USB4 and Thunderbolt cables
   Depends on: NETDEVICES [=y] && USB4 [=m] && INET [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])

---

USB4_NET is a replacement for THUNDERBOLT_NET and its setting reflects
that.

Cc: Jarod Wilson <jarod@redhat.com>