Add extra version bump to os-build branch
Currently, ark-latest is built, version bumped and the gitlab-ci.yml
file secretly copies that changelog back to os-build.

I think we can improve that process, but it requires a change in
the release behaviour, version bump os-build first, then create
ark-latest and version bump a second time.

The second version bump should have a small changelog and only
captures the 'extra' patches that are necessary for the release.

This _is_ a behaviour change for the changelog and kernel NVR.

Signed-off-by: Don Zickus <dzickus@redhat.com>