New configs in arch/powerpc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PPC_PROT_SAO_LPAR:

 This option adds support for PROT_SAO mappings from userspace
 inside LPARs on supported CPUs.

 This may cause issues when performing guest migration from
 a CPU that supports SAO to one that does not.

 If unsure, say N here.

 Symbol: PPC_PROT_SAO_LPAR [=n]
 Type  : bool
 Defined at arch/powerpc/Kconfig:863
   Prompt: Support PROT_SAO mappings in LPARs
   Depends on: PPC_BOOK3S_64 [=y]
   Location:
     -> Kernel options

Cc: kernel-patches@redhat.com