[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PCIE_HISI_ERR:

 Say Y here if you want error handling support
 for the PCIe controller's errors on HiSilicon HIP SoCs

 Symbol: PCIE_HISI_ERR [=n]
 Type  : bool
 Defined at drivers/pci/controller/Kconfig:301
   Prompt: HiSilicon HIP PCIe controller error handling driver
   Depends on: PCI [=y] && ACPI_APEI_GHES [=y] && (ARM64 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> PCI controller drivers

---

Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>