[redhat] New configs in drivers/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: AMD_PTDMA [=n]
 Type  : tristate
 Defined at drivers/dma/ptdma/Kconfig:2
   Prompt: AMD PassThru DMA Engine
   Depends on: DMADEVICES [=y] && X86_64 [=y] && PCI [=y]
   Location:
     -> Device Drivers
       -> DMA Engine support (DMADEVICES [=y])
 Selects: DMA_ENGINE [=y] && DMA_VIRTUAL_CHANNELS [=y]

---

 Symbol: INTEL_IDXD_COMPAT [=n]
 Type  : bool
 Defined at drivers/dma/Kconfig:299
   Prompt: Legacy behavior for idxd driver
   Depends on: DMADEVICES [=y] && PCI [=y] && X86_64 [=y]
   Location:
     -> Device Drivers
       -> DMA Engine support (DMADEVICES [=y])
 Selects: INTEL_IDXD_BUS [=m]

---

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>