redhat/self-test: Clean up data set
After running 'make dist-self-test-data' it was noticed that there have
been changes made to the kernel.spec file since the self-test
data was introduced and a few more variables need to be changed for
testing purposes.

Additional fix: Remove RHEL_RELEASE as this changes build-to-build.

Clean up the self-test data set.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>