[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: INTEL_TDX_GUEST [=n]
 Type  : bool
 Defined at arch/x86/Kconfig:881
   Prompt: Intel TDX (Trust Domain Extensions) - Guest Support
   Depends on: HYPERVISOR_GUEST [=y] && X86_64 [=y] && CPU_SUP_INTEL [=y] && X86_X2APIC [=y]
   Location:
     Main menu
       -> Processor type and features
         -> Linux guest support (HYPERVISOR_GUEST [=y])
 Selects: ARCH_HAS_CC_PLATFORM [=y] && X86_MEM_ENCRYPT [=y] && X86_MCE [=y]

---

Cc: David Arcari <darcari@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Dean Nelson <dnelson@redhat.com>
Cc: Lenny Szubowicz <lszubowi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
- CONFIG_INTEL_TDX_GUEST enabled.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>