[redhat] New configs in kernel/rcu
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: RCU_EXP_CPU_STALL_TIMEOUT [=0]
 Type  : integer
 Range : [0 21000]
 Defined at kernel/rcu/Kconfig.debug:85
   Prompt: Expedited RCU CPU stall timeout in milliseconds
   Depends on: RCU_STALL_COMMON [=y]
   Location:
     Main menu
       -> Kernel hacking
         -> RCU Debugging

---

Cc: Waiman Long <longman@redhat.com>
Cc: Artem Savkov <asavkov@redhat.com>
Cc: Clark Williams <williams@redhat.com>
Cc: Donghai Qiao <dqiao@redhat.com>
Cc: Phil Auld <pauld@redhat.com>
Cc: Wander Costa <wander@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>