[redhat] New configs in fs/nfsd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NFSD_V2 [=n]
 Type  : bool
 Defined at fs/nfsd/Kconfig:35
   Prompt: NFS server support for NFS version 2 (DEPRECATED)
   Depends on: NETWORK_FILESYSTEMS [=y] && NFSD [=m]
   Location:
     -> File systems
       -> Network File Systems (NETWORK_FILESYSTEMS [=y])
         -> NFS server support (NFSD [=m])
           -> NFS server support for NFS version 2 (DEPRECATED) (NFSD_V2 [=n])

---

Cc: Benjamin Coddington <bcodding@redhat.com>
Cc: Jeff Layton <jlayton@kernel.org>
Cc: Steve Dickson <steved@redhat.com>
Cc: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>