redhat:  adapt to upstream Makefile change
Upstream commit 67d7c3023a672c2 breaks our include of
Makefile.rhelver when the O= option is used (moves the
object file directory). Move the include of Makefile.rhelver
to a point where the absolute path to the source dir is
defined.

Thanks to Eder Zulian (ezulian) for tracking this down.

Signed-off-by: Clark Williams <williams@redhat.com>