Make Fedora debug configs more useful for debug
For years, we forced users to run debug kernels in Rawhide. As a result,
we kept some debug options turned off as they were a larger performance
impact.  The trade off was worth while, but as performce got worse with
some of the key debug options, we had to stop forcing users to the debug
kernels.  As we are no longer forcing them, let's make the debug kernel
a better actual debug kernel.  This brings us more inline with the RHEL
debug configs, though not quite identical.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>