[redhat] New configs in drivers/iio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ROHM_BU27034 [=n]
 Type  : tristate
 Defined at drivers/iio/light/Kconfig:292
   Prompt: ROHM BU27034 ambient light sensor
   Depends on: IIO [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Light sensors
           -> ROHM BU27034 ambient light sensor (ROHM_BU27034 [=n])
 Selects: REGMAP_I2C [=m] && IIO_GTS_HELPER [=n] && IIO_BUFFER [=y] && IIO_KFIFO_BUF [=m]

---

 Symbol: TI_ADS1100 [=n]
 Type  : tristate
 Defined at drivers/iio/adc/Kconfig:1232
   Prompt: Texas Instruments ADS1100 and ADS1000 ADC
   Depends on: IIO [=m] && I2C [=y]
   Location:
     -> Device Drivers
       -> Industrial I/O support (IIO [=m])
         -> Analog to digital converters
           -> Texas Instruments ADS1100 and ADS1000 ADC (TI_ADS1100 [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>