kernel.spec: drop obsolete ldconfig
This ldconfig is misplaced and over the years it was just
added to whatever sub-package %post section happened to be
the last one preceding it, e.g. zfcdump, kdump, xen, etc.

The weird side effect of this was a strange error for rtonly builds:
  /sbin/ldconfig: Can't create temporary cache file /etc/ld.so.cache~: Permission denied

This happened because all %post sections were behind %if conditions
that were never true. So instead of this ldconfig landing in %post somewhere
"randomly", it became part of commands for %install section and got
executed inside the build.

Remove it.

Signed-off-by: Jan Stancek <jstancek@redhat.com>