[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: SENSORS_HS3001 [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:737
   Prompt: Renesas HS3001 humidity and temperature sensors
   Depends on: HWMON [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
         -> Renesas HS3001 humidity and temperature sensors (SENSORS_HS3001 [=n])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>